Move test helper funcs into separate pkg
